requirements.txt
numpy
opencv-python
Pillow
matplotlib
scikit-learn
torch
torchvision
tqdm
pandas
watchdog
seaborn
scipy